Be more clever about updating the widgets.
Splits refresh out into setup and refresh. The button deleting magic is still
there but I can't test it.

Disable the slider if the governor is not "userspace". Also hide the label.
This makes the interface a bit clunky as the slider jumps about when the label
is hidden. The alternative is to have a timer check the current frequency so
that the label can be properly updated, but that's ugly.

There are now a lot of object variables. Should probably fix that...