Always show the slider and label. Fix a probable bug.
The user should always be able to use the slider without having to set the
governor first. Set the governor for them if they do this.

len(freq) should have been len(self.freqs). Mostly, these were equal anyway!

Also, try to do less in setup(), as it only has to be repeated in refresh().