The last bits of making slider work properly.
Disconnect and then reconnect the signal which updates the label if we're
setting the slider value because we have read it from the kernel. This
happens when the user changes governor, and we don't want to override their
change by setting the governor to userspace. It seems to be too difficult to
tell the difference between the user moving the slider and the code calling
setValue(), so I went for the simpler approach.

Make it slightly clearer that the slider doesn't show the actual frequency by
setting the label to "Dynamic" if the governor is not userspace, powersave or
performance.

Fix a bug where the governor could not be changed by moving the slider. The
file must be opened for writing, of course. :)